+++ sdk/attestation/attestation/recordings/node/attestationtokentests/recording_creatersasigningkey.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "3fc9b607de72d9c0c15ae925d17db4b5";
+module.exports.hash = "862057838898ead12506ac13eb6d25b9";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}