+++ sdk/attestation/attestation/recordings/node/tokencerttests/recording_getmetadataconfigshared.js
@@ -11,6 +11,8 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'no-store, no-cache',
   'Pragma',
   'no-cache',
+  'Content-Length',
+  '1317',
   'Content-Type',
   'application/json; charset=utf-8',
   'Expires',
@@ -22,28 +24,26 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'P3P',
   'CP="DSP CUR OTPi IND OTRi ONL FIN"',
   'x-ms-request-id',
-  'f8e6c954-2096-40c5-a8cd-a23c811c2400',
+  'e9e659b4-be46-42d8-9d97-53520ccf0400',
   'x-ms-ests-server',
-  '2.1.11774.11 - NCUS ProdSlices',
+  '2.1.11722.21 - SCUS ProdSlices',
   'Set-Cookie',
-  'fpc=AoeeOFCcvcNIrB8eRXdVKmq81LWYDAAAAD-EP9gOAAAA; expires=Thu, 24-Jun-2021 23:50:50 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'fpc=ApjvTYg-5uZEl0SYtSdv4Zy81LWYBwAAANujQNgOAAAA; expires=Fri, 25-Jun-2021 20:17:55 GMT; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
   'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
   'Set-Cookie',
   'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
   'Date',
-  'Tue, 25 May 2021 23:50:49 GMT',
-  'Content-Length',
-  '1317'
+  'Wed, 26 May 2021 20:17:54 GMT'
 ]);
 
-nock('https://shareduks.uks.attest.azure.net:443', {"encodedQueryParams":true})
+nock('https://sharedwus.wus.attest.azure.net:443', {"encodedQueryParams":true})
   .get('/.well-known/openid-configuration')
-  .reply(200, {"response_types_supported":["token","none"],"id_token_signing_alg_values_supported":["RS256"],"revocation_endpoint":"https://shareduks.uks.attest.azure.net/revoke","issuer":"https://shareduks.uks.attest.azure.net","jwks_uri":"https://shareduks.uks.attest.azure.net/certs","claims_supported":["cnf","nonce","x-ms-ver","x-ms-attestation-type","x-ms-policy-hash","x-ms-policy-signer","x-ms-sgx-is-debuggable","x-ms-sgx-mrenclave","x-ms-sgx-mrsigner","x-ms-sgx-svn","x-ms-sgx-ehd","x-ms-sgx-collateral","is-debuggable","sgx-mrsigner","sgx-mrenclave","product-id","svn","tee"]}, [
+  .reply(200, {"response_types_supported":["token","none"],"id_token_signing_alg_values_supported":["RS256"],"revocation_endpoint":"https://sharedwus.wus.attest.azure.net/revoke","issuer":"https://sharedwus.wus.attest.azure.net","jwks_uri":"https://sharedwus.wus.attest.azure.net/certs","claims_supported":["cnf","nonce","x-ms-ver","x-ms-attestation-type","x-ms-policy-hash","x-ms-policy-signer","x-ms-sgx-is-debuggable","x-ms-sgx-mrenclave","x-ms-sgx-mrsigner","x-ms-sgx-svn","x-ms-sgx-ehd","x-ms-sgx-collateral","is-debuggable","sgx-mrsigner","sgx-mrenclave","product-id","svn","tee"]}, [
   'Connection',
   'close',
   'Date',
-  'Tue, 25 May 2021 23:50:50 GMT',
+  'Wed, 26 May 2021 20:17:54 GMT',
   'Content-Type',
   'application/json; charset=utf-8',
   'Server',
@@ -51,7 +51,7 @@ nock('https://shareduks.uks.attest.azure.net:443', {"encodedQueryParams":true})
   'Content-Length',
   '573',
   'x-ms-request-id',
-  '00-4f6ccbaf1f559b2ec303fccb293bd236-0000000000000000-00',
+  '00-c14a1d8ade6a063e1df41afd934f6431-0000000000000000-00',
   'x-ms-maa-service-version',
   '1.10.01598.0001'
 ]);