+++ sdk/attestation/attestation/src/models/attestationResponse.ts
@@ -3,10 +3,8 @@
  * Licensed under the MIT License.
  *
  */
-
+import { HttpResponse } from "@azure/core-http";
 import { AttestationToken } from "./attestationToken";
-import { Serializer, Mapper, HttpResponse } from "@azure/core-http";
-
 /**
  * An AttestationResponse represents the response from the Microsoft Azure
  * Attestation service. It has two properties:
@@ -18,13 +16,9 @@ import { Serializer, Mapper, HttpResponse } from "@azure/core-http";
  */
  export class AttestationResponse<T>
  {
-     constructor(token: AttestationToken, serializer: Serializer, bodyMapper: Mapper, bodyTypeName: string, rawResult: any) {
+     constructor(token: AttestationToken, value: T, rawResult: any) {
         this.token = token;
-        this.value = serializer.deserialize(
-            bodyMapper,
-            token.get_body(),
-            bodyTypeName
-          );
+        this.value = value;
         this._response = rawResult._response;
      }