+++ sdk/attestation/attestation/test/utils/recordedClient.ts
@@ -6,15 +6,16 @@ import { Context } from "mocha";
 import { ClientSecretCredential } from "@azure/identity";
 import { env, Recorder, record, RecorderEnvironmentSetup } from "@azure/test-utils-recorder";
 
-import { AttestationClient, AttestationClientOptions } from "../../src/";
+import { AttestationClient, AttestationClientOptions, AttestationAdministrationClient } from "../../src/";
 import "./env";
 
 const replaceableVariables: { [k: string]: string } = {
   AZURE_CLIENT_ID: "azure_client_id",
   AZURE_CLIENT_SECRET: "azure_client_secret",
   AZURE_TENANT_ID: "azure_tenant_id",
-  ISOLATED_ATTESTATION_URL: "https://isolated_attestation_url",
-  AAD_ATTESTATION_URL: "https://aad_attestation_url",
+  ATTESTATION_LOCATION_SHORT_NAME: "wus",
+  ISOLATED_ATTESTATION_URL: "https://isolated_attestation_url.wus.attest.azure.net",
+  AAD_ATTESTATION_URL: "https://aad_attestation_url.wus.attest.azure.net",
   policySigningCertificate0: "policy_signing_certificate0",
   policySigningCertificate1: "policy_signing_certificate1",
   policySigningCertificate2: "policy_signing_certificate2",
@@ -44,7 +45,24 @@ export function createRecorder(context: Context): Recorder {
   return record(context, environmentSetup);
 }
 
-type EndpointType = "AAD" | "Isolated" | "Shared";
+export type EndpointType = "AAD" | "Isolated" | "Shared";
+
+export function getAttestationUri(endpointType: EndpointType) : string{
+  switch (endpointType) {
+    case "AAD": {
+      return  env.AAD_ATTESTATION_URL;
+    }
+    case "Isolated": {
+      return env.ISOLATED_ATTESTATION_URL;
+    }
+    case "Shared": {
+      return "https://shared" + env.ATTESTATION_LOCATION_SHORT_NAME + "." + env.ATTESTATION_LOCATION_SHORT_NAME + ".attest.azure.net";
+    }
+    default: {
+      throw new Error(`Unsupported endpoint type: ${endpointType}`);
+    }
+  }
+}
 
 export function createRecordedClient(
   endpointType: EndpointType,
@@ -55,18 +73,18 @@ export function createRecordedClient(
     env.AZURE_CLIENT_ID,
     env.AZURE_CLIENT_SECRET
   );
-  switch (endpointType) {
-    case "AAD": {
-      return new AttestationClient(credential, env.AAD_ATTESTATION_URL, options);
-    }
-    case "Isolated": {
-      return new AttestationClient(credential, env.ISOLATED_ATTESTATION_URL, options);
-    }
-    case "Shared": {
-      return new AttestationClient(credential, "https://shareduks.uks.attest.azure.net", options);
-    }
-    default: {
-      throw new Error(`Unsupported endpoint type: ${endpointType}`);
-    }
-  }
+
+  return new AttestationClient(credential, getAttestationUri(endpointType), options);
+}
+
+export function createRecordedAdminClient(
+  endpointType: EndpointType,
+  options?: AttestationClientOptions
+): AttestationAdministrationClient {
+  const credential = new ClientSecretCredential(
+    env.AZURE_TENANT_ID,
+    env.AZURE_CLIENT_ID,
+    env.AZURE_CLIENT_SECRET
+  );
+  return new AttestationAdministrationClient(credential, getAttestationUri(endpointType), options);
 }