+++ sdk/attestation/attestation/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Breaking Changes
 
 - Essentially completely rewritten. All existing functionality has been replaced.
+  - Removed `policy.get`, replaced with the `getPolicy` method of the new  `AttestationAdministrationClient` client object.
   - Removed `attestation.attestSgxEnclave`, `attestation.attestOpenEnclave`, `attestation.attestTpm`, and `attestation` property from attestationClient, replaced with `attestSgxEnclave`, `attestOpenEnclave` and `attestTpm`.
   - Removed `metadataConfiguration` and `signingCertificates` properties from attestationClient.
   - Removed `metadataConfiguration.get()` method, replaced with `client.getOpenIdMetadata()`.